+++ src/main/java/algorithms/curated170/medium/designahitcounter/DesignAHitCounter.java
@@ -1,4 +1,4 @@
-package algorithms.curated170.medium;
+package algorithms.curated170.medium.designahitcounter;
 
 import java.util.LinkedList;
 import java.util.List;
