+++ src/main/java/algorithms/curated170/medium/killprocess/KillProcessBFS.java
@@ -1,13 +1,13 @@
-package algorithms.curated170.medium;
+package algorithms.curated170.medium.killprocess;
 
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Queue;
 
-public class KillProcessRecursive {
-    public static List<Integer> killProcessRecursive(List<Integer> pid, List<Integer> ppid, int kill)
-    {
+public class KillProcessBFS {
+    public static List<Integer> killProcessBFS(List<Integer> pid, List<Integer> ppid, int kill) {
         HashMap<Integer, HashSet<Integer>> parentChildMap = new HashMap<>();
         for (int i = 0; i < pid.size(); i++) {
             int par = ppid.get(i);
@@ -24,24 +24,22 @@ public class KillProcessRecursive {
         }
 
         List<Integer> pTK = new LinkedList<>();
-        recursiveHelper(pTK, parentChildMap, kill);
-        return pTK;
-    }
-    private static void recursiveHelper(List<Integer> pTK, HashMap<Integer, HashSet<Integer>> parentChildMap, int kill)
-    {
-        pTK.add(kill);
-        if (parentChildMap.containsKey(kill)) {
-            for(int toKill : parentChildMap.get(kill))
-            {
-                recursiveHelper(pTK, parentChildMap, toKill);
+        Queue<Integer> toKill = new LinkedList<>();
+        toKill.offer(kill);
+        while (!toKill.isEmpty()) {
+            int killed = toKill.poll();
+            pTK.add(killed);
+            if (parentChildMap.containsKey(killed)) {
+                toKill.addAll(parentChildMap.get(killed));
             }
         }
+        return pTK;
     }
 
     public static void main(String[] args) {
         var pid = List.of(1, 3, 10, 5);
         var ppid = List.of(3, 0, 5, 3);
-        var k3 = killProcessRecursive(pid, ppid, 3);
-        System.out.println(k3);
+        var k2 = killProcessBFS(pid, ppid, 3);
+        System.out.println(k2);
     }
 }
