+++ src/main/java/algorithms/curated170/medium/killprocess/KillProcessDFS.java
@@ -1,13 +1,13 @@
-package algorithms.curated170.medium;
+package algorithms.curated170.medium.killprocess;
 
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Queue;
+import java.util.Stack;
 
-public class KillProcessBFS {
-    public static List<Integer> killProcessBFS(List<Integer> pid, List<Integer> ppid, int kill) {
+public class KillProcessDFS {
+    public static List<Integer> killProcessDFS(List<Integer> pid, List<Integer> ppid, int kill) {
         HashMap<Integer, HashSet<Integer>> parentChildMap = new HashMap<>();
         for (int i = 0; i < pid.size(); i++) {
             int par = ppid.get(i);
@@ -24,10 +24,10 @@ public class KillProcessBFS {
         }
 
         List<Integer> pTK = new LinkedList<>();
-        Queue<Integer> toKill = new LinkedList<>();
-        toKill.offer(kill);
-        while (!toKill.isEmpty()) {
-            int killed = toKill.poll();
+        Stack<Integer> toKill = new Stack<Integer>();
+        toKill.push(kill);
+        while (!toKill.empty()) {
+            int killed = toKill.pop();
             pTK.add(killed);
             if (parentChildMap.containsKey(killed)) {
                 toKill.addAll(parentChildMap.get(killed));
@@ -39,7 +39,7 @@ public class KillProcessBFS {
     public static void main(String[] args) {
         var pid = List.of(1, 3, 10, 5);
         var ppid = List.of(3, 0, 5, 3);
-        var k2 = killProcessBFS(pid, ppid, 3);
-        System.out.println(k2);
+        var k1 = killProcessDFS(pid, ppid, 3);
+        System.out.println(k1);
     }
 }
