+++ src/main/java/algorithms/curated170/medium/nestedlistweightsum2/NestedListWeightSum2.java
@@ -1,4 +1,4 @@
-package algorithms.curated170.medium;
+package algorithms.curated170.medium.nestedlistweightsum2;
 
 import java.util.List;
